removed some bugs, but does not work on startup

+++ alpha_drivers/include/alpha_drivers/RCOut_nodelet.h
@@ -15,7 +15,7 @@ namespace alpha_autopilot{
     virtual void onInit();
     ~RCOut();
     void rc_sub(const alpha_msgs::RCConstPtr msg);
-    void set_pulse(int16_t *pulse);
+    void set_pulse(int *pulse);
     Ada_ServoDriver pwm;
 
   };
